Changed some of the primitive types in the property class
Originally had the property price a double and changed it to an integer for consistency as well as making some of the GameController methods work.

+++ src/cis/monopoly/Player.java
@@ -133,11 +133,11 @@ public class Player {
         this.playPieceID = playPieceID;
     }
 
-    public void addBalance(double amount) {
+    public void addBalance(int amount) {
         playBalance += amount;
     }
 
-    public void subtractBalance(double amount) {
+    public void subtractBalance(int amount) {
         playBalance -= amount;
     }
 }
